+++ frontend/workflows/projectSelector/src/selector-reducer.tsx
@@ -1,102 +1,14 @@
-import * as React from "react";
 import type { clutch as IClutch } from "@clutch-sh/api";
 import _ from "lodash";
 
-import type { Action, ProjectState, State } from "./types";
+import {
+  deriveSwitchStatus,
+  exclusiveProjectDependencies,
+  PROJECT_TYPE_URL,
+  updateGroupstate,
+} from "./helpers";
+import type { Action, State } from "./types";
 import { Group } from "./types";
-
-const PROJECT_TYPE_URL = "type.googleapis.com/clutch.core.project.v1.Project";
-
-interface DependencyMappings {
-  upstreams?: { [dependency: string]: { [project: string]: boolean } };
-  downstreams?: { [dependency: string]: { [project: string]: boolean } };
-}
-
-const StateContext = React.createContext<State | undefined>(undefined);
-const useReducerState = () => {
-  return React.useContext(StateContext);
-};
-
-const DispatchContext = React.createContext<(action: Action) => void | undefined>(undefined);
-const useDispatch = () => {
-  return React.useContext(DispatchContext);
-};
-
-// TODO(perf): call with useMemo().
-const deriveSwitchStatus = (state: State, group: Group): boolean => {
-  return (
-    Object.keys(state[group]).length > 0 &&
-    Object.keys(state[group]).every(key => state[group][key].checked)
-  );
-};
-
-const updateGroupstate = (
-  state: State,
-  group: Group,
-  project: string,
-  projectState: ProjectState
-): State => {
-  const newState = { ...state };
-  if (project in newState[group]) {
-    // preserve the checked value if the project is already in the group
-    newState[group][project].checked = state[group][project].checked;
-    newState[group][project].custom = projectState.custom;
-  } else {
-    // we set the projectState to the default state passed in
-    newState[group][project] = projectState;
-  }
-
-  return newState;
-};
-
-const dependencyToProjects = (state: State, group: Group): DependencyMappings => {
-  const upstreamMap = {};
-  const downstreamMap = {};
-
-  const projects = Object.keys(state[group]);
-  projects.forEach(p => {
-    const { upstreams, downstreams } = state.projectData[p]?.dependencies;
-    upstreams[PROJECT_TYPE_URL]?.ids.forEach(u => {
-      if (!upstreamMap[u]) {
-        upstreamMap[u] = { [p]: true };
-      } else {
-        upstreamMap[u][p] = true;
-      }
-    });
-    downstreams[PROJECT_TYPE_URL]?.ids.forEach(d => {
-      if (!downstreamMap[d]) {
-        downstreamMap[d] = { [p]: true };
-      } else {
-        downstreamMap[d][p] = true;
-      }
-    });
-  });
-
-  return { upstreams: upstreamMap, downstreams: downstreamMap };
-};
-
-const exclusiveProjectDependencies = (
-  state: State,
-  group: Group,
-  project: string
-): { upstreams: string[]; downstreams: string[] } => {
-  const dependencyMap = dependencyToProjects(state, group);
-
-  const upstreams = [];
-  const downstreams = [];
-  _.forIn(dependencyMap.upstreams, (v, k) => {
-    if (v[project] && Object.keys(v).length === 1) {
-      upstreams.push(k);
-    }
-  });
-
-  _.forIn(dependencyMap.downstreams, (v, k) => {
-    if (v[project] && Object.keys(v).length === 1) {
-      downstreams.push(k);
-    }
-  });
-  return { upstreams, downstreams };
-};
 
 const selectorReducer = (state: State, action: Action): State => {
   switch (action.type) {
@@ -163,18 +75,23 @@ const selectorReducer = (state: State, action: Action): State => {
 
       // for the removed projects in Group.Projects we get their upstreams/dowstreams that are exclusive
       // (i.e. not shared by other projects that weren't removed)
-      const upstreams = [];
-      const downstreams = [];
+      const exclusiveUpstreams = [];
+      const exclusiveDownstreams = [];
       action.payload.projects.forEach(p => {
-        // get the upstreams/downstreams that are exclusive to the removed project(s)
-        const dependenices = exclusiveProjectDependencies(state, Group.PROJECTS, p);
-        upstreams.push(...dependenices.upstreams);
-        downstreams.push(...dependenices.downstreams);
+        /* get the upstreams/downstreams that are exclusive to the removed project(s)
+
+        TODO: (sperry) "remove_projects" supports bulk removal & since this func checks if upstreams/downstreams are exclusive
+        to a *single* project, if a dependency is shared by multiple *removed* projects, it wont't be removed. The func should
+        be refactored to take in the list of removed projects to account for this scenario.
+        */
+        const dependencies = exclusiveProjectDependencies(state, Group.PROJECTS, p);
+        exclusiveUpstreams.push(...dependencies.upstreams);
+        exclusiveDownstreams.push(...dependencies.downstreams);
       });
 
       // remove the upstreams/downstreams exclusive to the removed project(s)
-      newState[Group.UPSTREAM] = _.omit(state[Group.UPSTREAM], upstreams);
-      newState[Group.DOWNSTREAM] = _.omit(state[Group.DOWNSTREAM], downstreams);
+      newState[Group.UPSTREAM] = _.omit(state[Group.UPSTREAM], exclusiveUpstreams);
+      newState[Group.DOWNSTREAM] = _.omit(state[Group.DOWNSTREAM], exclusiveDownstreams);
 
       return newState;
     }
@@ -277,11 +194,4 @@ const selectorReducer = (state: State, action: Action): State => {
   }
 };
 
-export {
-  deriveSwitchStatus,
-  DispatchContext,
-  selectorReducer,
-  StateContext,
-  useDispatch,
-  useReducerState,
-};
+export default selectorReducer;
